+++ usr-local-cast-sbin/RMHS.sh
@@ -1,10 +1,15 @@
+#!/bin/bash
+
 #
 # Places Cast in Hotspot Mode
 #
 # PE1MSZ, PE1PLM, W0CHP
 #
 
-sudo systemctl stop mmdvmhost.service > /dev/null 2>&1
+# Don't run this if called from config page:
+if [ "$1" != "conf_page" ]; then
+    sudo systemctl stop mmdvmhost.service > /dev/null 2>&1
+fi
 
 sudo sed -i "/\[Modem\]/,/\[/ s/Hardware=.*$/Hardware=dvmpicasths/1" /etc/dstar-radio.mmdvmhost
 sudo sed -i "/\[Modem\]/,/\[/ s/Port=.*$/Port=\/dev\/ttyS2/1" /etc/mmdvmhost
@@ -18,5 +23,11 @@ sudo /usr/local/cast/bin/cast-reset
 sudo gpio mode 10 in
 sleep 1
 
-sudo systemctl start mmdvmhost.service > /dev/null 2>&1 &
+# Don't run this if called from config page:
+if [ "$1" != "conf_page" ]; then
+    sudo systemctl start mmdvmhost.service > /dev/null 2>&1 &
+fi
+
 sudo systemctl start castserial.service > /dev/null 2>&1 &
+
+exit 0
